+++ dal/src/main/java/catalogo/dal/FacturaDAO.java
@@ -1,5 +1,7 @@
 package catalogo.dal;
 
+import java.math.BigDecimal;
+
 import catalogo.tipos.Factura;
 import catalogo.tipos.Producto;
 
@@ -22,9 +24,9 @@ public interface FacturaDAO extends IpartekDAO {
 
 	public int insertFacturaProducto(int id_factura, int id_producto);
 
-	public Double getIvaTotal(int id);
+	public BigDecimal getIvaTotal(int id);
 	
-	public Double getPrecioTotal(int id);
+	public BigDecimal getPrecioTotal(int id);
 	
 	public int getMaxId();
 }
